feat(*): Add init function and split out admin code into separate class
